tests for some utils used in client gen as well
